requirements.txt
requests>=2.25.0
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.8.0
watchdog>=2.1.0
discord.py>=2.3.0 